Add OpenAI evaluation support helper
