+++ script/fill_call_ratings.php
@@ -2,7 +2,8 @@
 // fill_call_ratings.php
 // Analyse WisperTALK transcripts and populate rating fields in sales_call_ratings.
 
-require_once __DIR__ . '/../public_html/openai_evaluate.php';
+require_once __DIR__ . '/../app/Support/OpenAiEvaluate.php';
+require_once __DIR__ . '/db.php';
 
 // REM Use predefined OpenAI assistant unless overridden via OPENAI_ASSISTANT_ID
 $assistantId = getenv('OPENAI_ASSISTANT_ID') ?: 'asst_dxSC2TjWn45PX7JDdM8RpiyQ';
@@ -65,34 +66,6 @@ function process_missing_ratings(PDO $pdo, callable $evaluate): int
     }
     fwrite(STDERR, "Processed {$count} record(s)\n");
     return $count;
-}
-
-/**
- * Create a PDO connection using environment variables.
- *
- * @return PDO Database connection
- */
-function create_pdo_from_env(): PDO
-{
-    $connection = getenv('DB_CONNECTION') ?: 'mysql';
-    $database   = getenv('DB_DATABASE')   ?: 'salescallsanalyse';
-    $username   = getenv('DB_USERNAME')   ?: 'root';
-    $password   = getenv('DB_PASSWORD')   ?: '';
-    $host       = getenv('DB_HOST')       ?: '127.0.0.1';
-
-    if ($connection === 'sqlite') {
-        $dsn = 'sqlite:' . $database;
-        fwrite(STDERR, "Connecting using SQLite DSN {$dsn}\n");
-        return new PDO($dsn, null, null, [
-            PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION,
-        ]);
-    }
-
-    $dsn = "mysql:host={$host};dbname={$database};charset=utf8mb4";
-    fwrite(STDERR, "Connecting using MySQL DSN {$dsn}\n");
-    return new PDO($dsn, $username, $password, [
-        PDO::ATTR_ERRMODE => PDO::ERRMODE_EXCEPTION,
-    ]);
 }
 
 if (realpath(__FILE__) === realpath($_SERVER['SCRIPT_FILENAME'])) {
